requirements.txt
# DarkBottomLine Framework Requirements
# Core scientific packages
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
pandas>=1.3.0

# Physics analysis packages
awkward>=2.0.0
uproot>=5.0.0
correctionlib>=2.3.0

# Coffea and distributed computing
coffea>=2024.4.0
dask[complete]>=2022.0.0
distributed>=2022.0.0

# Machine learning
torch>=1.12.0
scikit-learn>=1.0.0

# Plotting and visualization
hist>=2.6.0
plotly>=5.0.0

# Data handling
pyarrow>=8.0.0

# Development and testing
pytest>=6.0.0
jupyter>=1.0.0
ipykernel>=6.0.0

# Configuration
pyyaml>=6.0

# Optional: Advanced plotting
seaborn>=0.11.0

# Optional: Progress bars
tqdm>=4.64.0

# Optional: Memory profiling
memory-profiler>=0.60.0